requirements.txt
asgiref==3.2.3
Django==3.0.2
django-markdownx==3.0.1
Markdown==2.6.11
Pillow==7.0.0
py-gfm==0.1.4
pytz==2019.3
sqlparse==0.3.0
